the clouds appear now, but in a straight line wtf lmao

+++ src/main/java/com/qendolin/betterclouds/clouds/Sampler.java
@@ -55,7 +55,7 @@ public class Sampler {
         return hashToFloat(x, z, 'Z');
     }
 
-    public float sample(int x, int z, float cloudiness, float fuzziness, float scale) {
+    public float sample(int x, int z, double cloudiness, double fuzziness, double scale) {
         // TODO: A vanilla like cloud distribution is not possible with this function
         double value = NOISE.sample(x / scale / 128f, z / scale / 128f, false);
         value = value / 2 + 0.5;
